# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Tableros Kanban",
        "description": "API para la gestión de tableros, columnas y tareas tipo Kanban. Permite a los usuarios crear tableros personalizados, definir columnas (estados) y gestionar tareas de manera dinámica. Todas las rutas están protegidas por autenticación y auditoría.",
        "contact": {
            "name": "Soporte API",
            "email": "marcelod.ferreira.dev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API principal"
        }
    ],
    "paths": {
...
        "/api/boards/{boardId}/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Listar tareas de un tablero",
                "operationId": "39eb72ccb28bb9a4d8c162061ddc4052",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "board_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "column_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Investigar tema 1"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Buscar información sobre el tema 1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-07-08T12:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-07-08T12:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Tareas obtenidas correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear una tarea en un tablero",
                "operationId": "54f57fb7785b16cec7e3683f04c43d10",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "column_id"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Investigar tema 1"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Buscar información sobre el tema 1"
                                    },
                                    "column_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "board_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "column_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "titulo": {
                                                    "type": "string",
                                                    "example": "Investigar tema 1"
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "Buscar información sobre el tema 1"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-07-08T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-07-08T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Tarea creada exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/boards/{boardId}/tasks/{taskId}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualizar una tarea",
                "operationId": "928c06a813ce641c7198e10ad32b9fe2",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Nuevo título"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Nueva descripción"
                                    },
                                    "column_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "board_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "column_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "titulo": {
                                                    "type": "string",
                                                    "example": "Nuevo título"
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "Nueva descripción"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-07-08T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-07-08T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Tarea actualizada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "dd0fd06279769938a3c133510bf15957",
                "parameters": [
                    {
                        "name": "boardId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Tarea eliminada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}